Corrections Create Offer and Add Create SubscribeOffer

+++ AMAP_FARM/AMAP_FARM/Models/OfferItem.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public required int OfferId { get; set; }
         public required int ProductId { get; set; }
-        public required decimal Price { get; set; }
-        public required int SaleQuantity { get; set; }
-        public int SoldQuantity { get; set; }
+        public required float Price { get; set; }
+        public required float SaleQuantity { get; set; }
+        public float SoldQuantity { get; set; }
     }
 }
